requirements.txt
Flask
qrcode
pillow
gunicorn
openpyxl
werkzeug
psycopg2-binary